% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopticulate.R
\docType{package}
\name{hopticulate}
\alias{hopticulate}
\alias{hyperopt}
\alias{fmin}
\alias{space_eval}
\alias{hopticulate-package}
\title{Hyperopt via reticulate}
\description{
`hyperopt` is a popular python package for automated hyperparameter
tuning.
}
\details{
This package, `hopticulate`, is a minimal wrapper for using hyperopt
through the most excellent reticulate package.
}
\examples{
fmin(function(x) (x - 1)**2, hp$normal("mu", 0, 1), algo=tpe$suggest, max_evals=10)

}
\references{
\url{https://github.com/hyperopt/hyperopt}
}
