% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopticulate.R
\name{search-space}
\alias{search-space}
\alias{hp}
\alias{hp.choice}
\alias{hp.pchoice}
\alias{hp.randint}
\alias{hp.uniformint}
\alias{hp.normal}
\alias{hp.lognormal}
\alias{hp.qnormal}
\alias{hp.qlognormal}
\alias{hp.uniform}
\alias{hp.loguniform}
\alias{hp.quniform}
\alias{hp.qloguniform}
\title{Search Space Definitions}
\description{
Declare a search space using the following:
}
\details{
\itemize{

\item Discrete choices
  \itemize{
    \item \code{hp.choice}
    \item \code{hp.pchoice}
  }

\item Integers
  \itemize{
    \item \code{hp.randint}
    \item \code{hp.uniformint}
  }

\item Normal
  \itemize{
    \item \code{hp.normal}
    \item \code{hp.lognormal}
    \item \code{hp.qnormal}
    \item \code{hp.qlognormal}
  }

\item Uniform
  \itemize{
    \item \code{hp.uniform}
    \item \code{hp.loguniform}
    \item \code{hp.quniform}
    \item \code{hp.qloguniform}
  }

}
}
