% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopticulate.R
\docType{package}
\name{hopticulate}
\alias{hopticulate}
\alias{hyperopt}
\alias{fmin}
\alias{space_eval}
\alias{Trial}
\alias{pyll.stochastic.sample}
\alias{hopticulate-package}
\title{Hyperopt via reticulate}
\description{
\code{hyperopt} is a popular python package for automated hyperparameter
tuning.
}
\details{
This package, \code{hopticulate}, is a minimal wrapper for using \code{hyperopt}
through the most excellent \code{reticulate} package.
}
\examples{
\dontshow{ if(!is.null(fmin))   }
  fmin(function(x) (x - 1)**2, hp.normal("mu", 0, 1), algo=tpe.suggest, max_evals=10)

}
\references{
\url{https://github.com/hyperopt/hyperopt}
}
