% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopticulate.R
\docType{package}
\name{hopticulate}
\alias{hopticulate}
\alias{hyperopt}
\alias{hopticulate-package}
\title{Hyperopt via reticulate}
\description{
`hyperopt` is a popular python package for automated hyperparameter
tuning.
}
\details{
This package, `hopticulate`, is a minimal wrapper for using hyperopt
through the most excellent reticulate package.
}
\references{
\url{https://github.com/hyperopt/hyperopt}
}
