% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hopticulate.R
\name{search-space}
\alias{search-space}
\alias{hp}
\alias{hp.choice}
\alias{hp.pchoice}
\alias{hp.randint}
\alias{hp.uniformint}
\alias{hp.normal}
\alias{hp.lognormal}
\alias{hp.qnormal}
\alias{hp.qlognormal}
\alias{hp.uniform}
\alias{hp.loguniform}
\alias{hp.quniform}
\alias{hp.qloguniform}
\title{Search Space Definitions}
\description{
Search Space Definitions
}
